Multiple output nodes, change input format
MLP will now function with multiple output nodes.
The MLP now accepts data in array format, this will make the MLP work with more general problems (previous format was designed for Assignment 1), this new format also means the MLP is not dependant on the PatternLoader class  which can now be removed.
The Client class is now responsible for loading in data (no Pattern Loader), and is currently testing the MLP with the XOR problem, which has had its data files reformatted to work with the new input format.